requirements.txt
sounddevice==0.5.1
webrtcvad==2.0.10
dotenv==0.9.9
simpleaudio==1.0.4
pydub==0.25.1
websockets==15.0
openwakeword==0.6.0
tensorflow==2.18.0
pyaudio==0.2.14
psycopg2-binary==2.9.10
fastapi==0.115.9
uvicorn==0.34.0
python-multipart==0.0.20
pandas==2.2.3
pyyaml==6.0.2
openpyxl
openai==1.72.0
paho-mqtt==2.1.0
zeroconf==0.132.2
netifaces==0.11.0
aiohttp==3.9.1